Issue #18423: Add SCR tracing and move test to unique server/class
