+++ dev/com.ibm.ws.webcontainer.servlet.4.0_fat/publish/servers/servlet40_sameSiteSecurity/bootstrap.properties
@@ -1,0 +1,16 @@
+###############################################################################
+# Copyright (c) 2021 IBM Corporation and others.
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the Eclipse Public License v1.0
+# which accompanies this distribution, and is available at
+# http://www.eclipse.org/legal/epl-v10.html
+#
+# Contributors:
+#     IBM Corporation - initial API and implementation
+###############################################################################
+bootstrap.include=../testports.properties
+osgi.console=7777
+com.ibm.ws.logging.trace.specification=*=info=enabled:HTTPChannel=all:TCPChannel=all:GenericBNF=all:logservice=all
+com.ibm.ws.logging.max.file.size=20
+com.ibm.ws.logging.max.files=10
+com.ibm.ws.classloading.tcclLockWaitTimeMillis=60000
